Basic chrome debugger support
