+++ src/reanimated2/PlatformChecker.ts
@@ -1,0 +1,21 @@
+import { Platform } from 'react-native';
+
+export function isJest(): boolean {
+  return !!process.env.JEST_WORKER_ID;
+}
+
+export function isChromeDebugger(): boolean {
+  return !(global as any).nativeCallSyncHook;
+}
+
+export function isWeb(): boolean {
+  return Platform.OS === 'web';
+}
+
+export function shouldBeUseWeb() {
+  return isJest() || isChromeDebugger() || isWeb();
+}
+
+export function nativeShouldBeMock() {
+  return isJest() || isChromeDebugger();
+}
