+++ src/reanimated2/core.ts
@@ -4,6 +4,7 @@
 import NativeReanimated from './NativeReanimated';
 import { Platform } from 'react-native';
 import { addWhitelistedNativeProps } from '../ConfigHelper';
+import { nativeShouldBeMock } from './PlatformChecker';
 
 global.__reanimatedWorkletInit = function (worklet) {
   worklet.__worklet = true;
@@ -21,7 +22,7 @@ const testWorklet = () => {
 };
 
 export const checkPluginState = (throwError = true) => {
-  if (!testWorklet.__workletHash && !process.env.JEST_WORKER_ID) {
+  if (!testWorklet.__workletHash && !shouldBeUseWeb()) {
     if (throwError) {
       throw new Error(
         "Reanimated 2 failed to create a worklet, maybe you forgot to add Reanimated's babel plugin?"
@@ -122,9 +123,11 @@ export function getViewProp(viewTag, propName) {
 }
 
 let _getTimestamp;
-if (process.env.JEST_WORKER_ID) {
+console.log('NativeReanimated', NativeReanimated);
+if (nativeShouldBeMock()) {
   _getTimestamp = () => {
-    return Date.now();
+    console.log('NativeReanimated', NativeReanimated);
+    return NativeReanimated.getTimestamp();
   };
 } else {
   _getTimestamp = () => {
